make app read-only
