+++ src/app/album/[key]/page.tsx
@@ -5,10 +5,8 @@ import type { Metadata, ResolvingMetadata } from "next";
 import { cache } from "react";
 import { findRelatedItems, getSourceItemByKey } from "@/util/services";
 import { Album } from "@/util/services/type";
-import { services } from "@/components/SupportedServices";
-import { Share } from "@/components/Share";
 import { ServiceLogo } from "@/components/ServiceLogo";
-import { HomeLink } from "@/components/HomeLink";
+import { services } from "@/util/services/list";
 
 const fallbackCover = "/img/cover-fallback.png";
 
@@ -97,7 +95,6 @@ export default async function Page({ params }: Props) {
 
   return (
     <main className="min-h-full mx-auto max-w-2xl mt-2 md:mt-6 px-2">
-      <HomeLink />
       <Image
         className="rounded-md drop-shadow-md"
         src={cover}
@@ -129,8 +126,6 @@ export default async function Page({ params }: Props) {
           );
         })}
       </ol>
-
-      <Share />
     </main>
   );
 }
