+++ src/util/services/list.ts
@@ -1,0 +1,34 @@
+export const services = [
+  {
+    key: "spotify",
+    name: "Spotify",
+    logo: "/img/logo/spotify.svg",
+    example: "https://open.spotify.com/track/4PTG3Z6ehGkBFwjybzWkR8",
+  },
+  {
+    key: "applemusic",
+    name: "Apple Music",
+    logo: "/img/logo/appleMusic.svg",
+    example:
+      "https://music.apple.com/fr/album/never-gonna-give-you-up/1559523357?i=1559523359",
+  },
+  {
+    key: "deezer",
+    name: "Deezer",
+    logo: "/img/logo/deezer.svg",
+    example: "https://www.deezer.com/us/track/14408104",
+  },
+  {
+    key: "Tidal",
+    name: "Tidal",
+    logo: "/img/logo/Tidal.svg",
+    example: "https://tidal.com/browse/track/101982419",
+  },
+  {
+    key: "youtubemusic",
+    name: "Youtube Music",
+    logo: "/img/logo/YoutubeMusic.svg",
+    example: "https://music.youtube.com/watch?v=lYBUbBu4W08",
+    soon: false,
+  },
+];
