+++ src/util/services/spotify/index.ts
@@ -7,72 +7,6 @@ import { ResourceType } from "@/util/validators/type";
 class Spotify {
   public static _name: "spotify";
 
-  static async search(type: ResourceType, params: SearchParams, market = "US") {
-    return await backOff(() => Spotify._search(type, params, market), {
-      numOfAttempts: 5,
-      maxDelay: 3500,
-    });
-  }
-
-  static async _search(
-    type: ResourceType,
-    params: SearchParams,
-    market = "US"
-  ): Promise<SearchResult | null> {
-    const token = await this.getAuthToken();
-    const q = new URLSearchParams();
-    let query = "";
-    if (params.track_name) query += params.track_name;
-    if (params.artist_name) query += " " + params.artist_name;
-    if (params.album_name) query += " " + params.album_name;
-    q.set("q", query);
-    q.set("type", type);
-    q.set("market", market);
-    q.set("limit", "1");
-    q.set("include_external", "audio");
-
-    const response = await fetch(
-      `https://api.spotify.com/v1/search?${q.toString()}`,
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      }
-    );
-    if (response.status === 401) {
-      const message = await response.text();
-      console.log("response", response.status, message);
-      return await this.deleteAuthToken();
-    }
-
-    const result = await response.json();
-
-    if (type === "artist" && result?.artists?.items.length > 0) {
-      const item = result.artists.items[0];
-      const key = this.getKey(item.id, "artist");
-      const value = this.parseArtist(key, item);
-      await kv.set(key, JSON.stringify(value));
-      return value;
-    }
-
-    if (type === "album" && result?.albums?.items?.length > 0) {
-      const item = result.albums.items[0];
-      const key = this.getKey(item.id, "album");
-      const value = this.parseAlbum(key, item);
-      await kv.set(key, JSON.stringify(value));
-      return value;
-    }
-    if (type === "track" && result?.tracks?.items?.length > 0) {
-      const item = result.tracks.items[0];
-      const key = this.getKey(item.id, "track");
-      const value = this.parseTrack(key, item);
-      await kv.set(key, JSON.stringify(value));
-      return value;
-    }
-
-    return null;
-  }
-
   private static async deleteAuthToken() {
     log.error("Spotify token expired");
     await kv.rename("spotify_access_token", "spotify_access_token_old");
