+++ src/util/services/youtubeMusic/index.ts
@@ -7,62 +7,6 @@ import { ResourceType } from "@/util/validators/type";
 class YoutubeMusic {
   public static _name: "youtubeMusic";
 
-  static async search(type: ResourceType, params: SearchParams, market = "US") {
-    return await backOff(() => YoutubeMusic._search(type, params, market), {
-      numOfAttempts: 5,
-      maxDelay: 3500,
-    });
-  }
-
-  static async _search(
-    type: ResourceType,
-    params: SearchParams,
-    market = "US"
-  ): Promise<SearchResult | null> {
-    let query = "";
-    if (params.track_name) query += params.track_name;
-    if (params.artist_name) query += " " + params.artist_name;
-    if (params.album_name) query += " " + params.album_name;
-    const api = new YTMusic();
-    await api.initialize();
-
-    if (type === "artist") {
-      const results = await api.searchArtists(query);
-      const item = results[0];
-      if (!item) return null;
-
-      const key = this.getKey(item.artistId, type);
-      const value = this.parseArtist(key, item);
-      await kv.set(key, JSON.stringify(value));
-      return value;
-    }
-
-    if (type === "album") {
-      const results = await api.searchAlbums(query);
-      const item = results[0];
-      if (!item) return null;
-
-      const key = this.getKey(item.albumId, type);
-
-      const value = this.parseAlbum(key, item);
-      await kv.set(key, JSON.stringify(value));
-      return value;
-    }
-
-    if (type === "track") {
-      const results = await api.searchSongs(query);
-      const item = results[0];
-      if (!item) return null;
-
-      const key = this.getKey(item.videoId, type);
-      const value = this.parseTrack(key, item);
-      await kv.set(key, JSON.stringify(value));
-      return value;
-    }
-
-    return null;
-  }
-
   static getKey(id: string, type: ResourceType) {
     return `youtubeMusic_${type}_${id}`;
   }
