--- src/util/validators/common.ts
@@ -1,1 +1,0 @@
-export const NOT_VALID_LINK = "Not a valid link";
